[FIX] Nâng cấp UI Trang Chủ

+++ SaleCar/src/main/java/code/salecar/service/product/DiscountService.java
@@ -48,7 +48,7 @@ public class DiscountService {
                 productIds.addAll(productDAO.getProductIdsByBrandId((int) discount.getEntityId()));
                 break;
             case CATEGORY:
-                productIds.addAll(productDAO.getProductIdsByCategoryId((int) discount.getEntityId()));
+                productIds.addAll(productDAO.getProductIdsByCategoryId( discount.getEntityId()));
                 break;
         }
